use double for percentage

+++ src/main/scala/peregin/gpv/gui/video/SimpleVideoPlayer.scala
@@ -13,13 +13,13 @@ import scala.concurrent._
 
 trait SimpleVideoPlayerFactory extends VideoPlayerFactory {
   override def createPlayer(url: String, telemetry: Telemetry, imageHandler: Image => Unit,
-                            shiftHandler: => Long, timeUpdater: (Long, Int) => Unit) =
+                            shiftHandler: => Long, timeUpdater: (Long, Double) => Unit) =
     new SimpleVideoPlayer(url, telemetry, imageHandler, shiftHandler, timeUpdater)
 }
 
 class SimpleVideoPlayer(url: String, telemetry: Telemetry,
                   imageHandler: Image => Unit, shiftHandler: => Long,
-                  timeUpdater: (Long, Int) => Unit) extends VideoPlayer with Logging {
+                  timeUpdater: (Long, Double) => Unit) extends VideoPlayer with Logging {
 
   @volatile var running = true
